% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loglikBM}
\alias{loglikBM}
\title{Calculate log likelihood using Kalman Filter}
\usage{
loglikBM(par_t, yt, days, N = 288, P = 5)
}
\description{
This function returns the average log likehood value of the stochastic periodicity model, given the input parameters.
}

