% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{spotvol}
\alias{spotVol}
\alias{spotvol}
\alias{spotvol-package}
\title{spotvol}
\usage{
spotVol(pdata, dailyvol = "bipower", periodicvol = "TML", on = "minutes",
  k = 5, dummies = FALSE, P1 = 4, P2 = 2, marketopen = "09:30:00",
  marketclose = "16:00:00")
}
\description{
Estimating Spot Volatility
}
\details{
Spot volatility estimates are important for highfrequency trading signal generation and risk management.
Several spot volatility estimators have been proposed over the last years.
This project aims at implementing them in the R package highfrequency.
}

