% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ssmodel}
\alias{ssmodel}
\title{Generate state space model}
\usage{
ssmodel(par_t, days, N = 288, P1 = 5, P2 = 5)
}
\description{
This function creates the state space matrices from the input parameters.
The output is in the format used by the FKF package.
}

