% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stochper}
\alias{stochper}
\title{Stochastic periodicity model}
\usage{
stochper(mR, init = list(), P1 = 5, P2 = 5)
}
\description{
This function estimates the spot volatility by using the stochastic periodcity model of Beltratti & Morana (2001)
}

