% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spotvol}
\alias{spotvol}
\title{spotvol}
\usage{
spotvol(pdata, dailyvol = "bipower", periodicvol = "TML", on = "minutes",
  k = 5, dummies = FALSE, P1 = 5, P2 = 5, marketopen = "09:30:00",
  marketclose = "16:00:00", method = "detper", init = list())
}
\description{
Estimate spot volatility
}
\examples{
data(sample)
vol1 <- spotvol(sample)
init = list(sigma = 0.03, sigma_mu = 0.005, sigma_h = 0.007,
   sigma_k = 0.06, phi = 0.194, rho = 0.986, mu = c(1.87,-0.42),
   delta_c = c(0.25, -0.05, -0.2, 0.13, 0.02), delta_s = c(-1.2, 0.11, 0.26, -0.03, 0.08))
vol2 <- spotvol(sample, method = "stochper", init = init)
plot(vol1$spotvol, type="l")
lines(vol2$spotvol, col="red")
}

