% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spotvol}
\alias{spotvol}
\title{spotvol}
\usage{
spotvol(data, makeReturns = TRUE, method = "detper", on = "minutes",
  k = 5, marketopen = "09:30:00", marketclose = "16:00:00", ...)
}
\arguments{
\item{data}{\code{xts} object, containing a price or return series. If the data consists
of returns, set \code{makeReturns} to \code{FALSE}.}

\item{makeReturns}{boolean, if \code{TRUE} the function will calculate returns from the price
series \code{data}. If \code{data} already consists of returns, set to \code{FALSE}.}

\item{method}{specifies which method will be used to estimate the spot volatility. Options
include \code{"detper"} and \code{"stochper"}. See Details.}

\item{on}{string indicating the time scale in which \code{k} is expressed.
Possible values are: \code{"secs", "seconds", "mins", "minutes", "hours"}.}

\item{k}{positive integer, indicating the number of periods to aggregate over.
E.g. to aggregate an \code{xts} object to the 5 minute frequency, set \code{k = 5} and
\code{on = "minutes"}.}

\item{marketopen}{the market opening time. By default, \code{marketopen = "09:30:00"}.}

\item{marketclose}{the market closing time. By default, \code{marketclose = "16:00:00"}.}

\item{...}{method-specific parameters (see 'Details').}
}
\value{
A \code{spotvol} object, which is a list containing one or more of the following outputs,
depending on the method used:

\code{spot}

An \code{xts} object containing spot volatility estimates \eqn{\sigma_{t,i}}, reported for each interval
\eqn{i} between \code{marketopen} and \code{marketclose} for every day \eqn{t} in \code{data}. The length
of the intervals is specifiedby \code{k} and \code{on}. Methods that provide this output: All.

\code{daily}

An \code{xts} object containing estimates of the daily volatility levels for each day \eqn{t} in \code{data},
if the used method decomposed spot volatility into a daily and an intraday component. Methods that provide this
output: \code{"detper"}.

\code{periodic}

Estimates of the intraday periodicity factor for each day interval \eqn{i} between \code{marketopen}
and \code{marketclose}, if the spot volatility was decomposed into a daily and an intraday component.
Methods that provide this output: \code{"detper"}.

\code{par}

A named list containing parameter estimates, for methods that estimate a parametric model.
Methods that provide this output: \code{"stochper"}.
}
\description{
The \code{spotvol} function offers several methods to estimate spot volatility
and its intraday seasonality, using high-frequency data. It returns an object of
class \code{spotvol}, which can contain various outputs, depending on the method
used. See 'Details' for a description of each method. In any case, the output will
contain the spot volatility estimates.

The input can consist of price data or return data, either tick by tick or sampled at
set intervals. The data will be converted to equispaced high-frequency returns \eqn{r_{t,i}}
(read: the \eqn{i}th return on day \eqn{t}).
}
\details{
The following estimation methods can be specified in \code{method}:

\strong{Deterministic periodicity method (\code{"detper"})}

Parameters:
\tabular{ll}{
\code{dailyvol} \tab A string specifying the estimation method for the daily component
\eqn{s_t}. Possible values are \code{"bipower", "rv", "medrv"}. Default = \code{"bipower"}. \cr
\code{periodicvol} \tab A string specifying the estimation method for the component of intraday
volatility, that depends in a deterministic way on the intraday time at which the return is observed.
Possible values are \code{"TML", "SD", "WSD", "OLS"}. See Boudt et al. (2011) for details. Default = \code{"TML"}.\cr
\code{P1} \tab A positive integer corresponding to the number of cosinus terms
 used in the flexible Fourier specification of the periodicity function,
 see Andersen et al. (1997) for details. Default = 5. \cr
\code{P2} \tab Same as \code{P1}, but for the sinus terms. Default = 5.\cr
\code{dummies} \tab Boolean: in case it is \code{TRUE}, the parametric estimator of periodic standard
deviation specifies the periodicity function as the sum of dummy variables corresponding to each intraday period.
If it false, the parametric estimator uses the flexible Fourier specification. Default = \code{FALSE}.
}
Outputs (see 'Value' for a full description of each component):
\itemize{
\item{\code{spot}}
\item{\code{daily}}
\item{\code{periodic}}
}
The spot volatility is decomposed into a deterministic periodic factor \eqn{f_{i}}
(identical for every day in the sample) and a daily factor \eqn{s_{t}} (identical for all
observations within a day). Both components are then estimated separately. For more
details, see Taylor and Xu (1997) and Andersen and Bollerslev (1997). The jump robust
versions by Boudt et al. (2011) have also been implemented.

\strong{Stochastic periodicity method (\code{"stochper"})}

Parameters:
\tabular{ll}{
\code{P1} \tab A positive integer corresponding to the number of cosinus terms
 used in the flexible Fourier specification of the periodicity function.
 Default = 5. \cr
\code{P2} \tab Same as \code{P1}, but for the sinus terms. Default = 5.\cr
\code{init} \tab A named list of initial values to be used in the optimization routine (\code{"BFGS"}
in \code{optim}). Default = \code{list(sigma = 0.03, sigma_mu = 0.005, sigma_h = 0.005, sigma_k = 0.05,
phi = 0.2, rho = 0.98, mu = c(2, -0.5), delta_c = rep(0, max(1,P1)), delta_s = rep(0, max(1,P2)))}.
See Beltratti & Morana (2001) for a definition of each parameter. \code{init} can contain any number of
these parameters. For parameters not specified in \code{init}, the default initial value will be used.
}
Outputs (see 'Value' for a full description of each component):
\itemize{
\item{\code{spot}}
\item{\code{par}}
}
This method by Beltratti and Morana (2001) assumes the periodicity factor to be stochastic.
The spot volatility estimation is split into four components: a random walk, an autoregressive
process, a stochastic cyclical process and a deterministic cyclical process. The model is
estimated using a quasi-maximum likelihood method based on the Kalman Filter. The package
\code{FKF} is used to apply the Kalman filter.
}
\examples{
data(sample_prices_5min)
vol1 <- spotvol(sample_prices_5min)
init = list(sigma = 0.03, sigma_mu = 0.005, sigma_h = 0.007,
   sigma_k = 0.06, phi = 0.194, rho = 0.986, mu = c(1.87,-0.42),
   delta_c = c(0.25, -0.05, -0.2, 0.13, 0.02), delta_s = c(-1.2, 0.11, 0.26, -0.03, 0.08))
vol2 <- spotvol(sample_prices_5min, method = "stochper", init = init)
vol3 <- spotvol(sample_prices_5min, method = "kernel")
plot(vol1$spot, type="l")
lines(vol2$spot, col="red")
lines(vol3$spot, col="blue")
legend("topright", c("detper", "stochper", "kernel"), col = c("black", "red", "blue"), lty=1, inset=0.02, bg="white")
}
\references{
Andersen, T. G. and T. Bollerslev (1997). Intraday periodicity and volatility
persistence in financial markets. Journal of Empirical Finance 4, 115-158.

Beltratti, A. and C. Morana (2001). Deterministic and stochastic methods for estimation
of intraday seasonal components with high frequency data. Economic Notes 30, 205-234.

Boudt K., Croux C. and Laurent S. (2011). Robust estimation of intraweek periodicity
in volatility and jump detection. Journal of Empirical Finance 18, 353-367.

Taylor, S. J. and X. Xu (1997). The incremental volatility information in one million
foreign exchange quotations. Journal of Empirical Finance 4, 317-340.
}

